refactor: extrai logica de requisicao de dados para custom hook
